+++ assets/js/app.js
@@ -1,22 +1,18 @@
-// Import dependencies
 /* global $ */
-require( 'datatables.net-bs4' )();
-require( 'datatables.net-select' )( window, $ );
+require('datatables.net-bs4')();
+require('datatables.net-select')(window, $);
 
-import { Socket } from "phoenix"
-import "phoenix_html"
-import Turbolinks from "turbolinks"
+import 'phoenix_html';
+import Turbolinks from 'turbolinks';
 
 // Import local files
-import { setTimeagoStrings } from './utils/functions';
+import { configPublicChannel } from './socket';
+import { configTimeago } from './utils/functions';
 import loadView from './views/loader';
 
 function handleDOMContentLoaded() {
-  // Get the current view name
-  const viewName = $('body').data('js-view-name');
-
   // Load view class and mount it
-  const ViewClass = loadView(viewName);
+  const ViewClass = loadView($('body').data('js-view-name'));
   const view = new ViewClass();
   view.mount();
 
@@ -27,16 +23,13 @@ function handleDocumentUnload() {
   window.currentView.unmount();
 }
 
-document.addEventListener("turbolinks:load", handleDOMContentLoaded)
-document.addEventListener("turbolinks:before-visit", handleDocumentUnload)
+// ALL FOLLOWING FUNCTIONS WILL BE EXECUTED ONLY ONCE (ON THE INITIAL PAGE LOAD)
 
-// set timeago local stings
-setTimeagoStrings()
+// Configuration
+configTimeago();
+configPublicChannel();
 
 // initialize TurboLinks
-Turbolinks.start()
-
-// connect socket
-const socket = new Socket("/socket", { params: { token: $('meta[name="token"]').attr("content") }})
-socket.connect()
-export default socket
+document.addEventListener('turbolinks:load', handleDOMContentLoaded);
+document.addEventListener('turbolinks:before-visit', handleDocumentUnload);
+Turbolinks.start();