+++ assets/js/utils/functions.js
@@ -1,13 +1,34 @@
 /* global tinymce, jQuery */
+export function configTimeago() {
+  jQuery.timeago.settings.strings = {
+    prefixAgo: null,
+    prefixFromNow: null,
+    suffixAgo: 'ago',
+    suffixFromNow: 'from now',
+    seconds: 'less than a minute',
+    minute: 'a minute',
+    minutes: '%d minutes',
+    hour: 'an hour',
+    hours: '%d hours',
+    day: 'a day',
+    days: '%d days',
+    month: 'a month',
+    months: '%d months',
+    year: 'a year',
+    years: '%d years',
+    wordSeparator: ' ',
+    numbers: []
+  };
+}
 
-export function string_to_slug (str) {
+export function string_to_slug(str) {
   str = str.replace(/^\s+|\s+$/g, ''); // trim
   str = str.toLowerCase();
 
   // remove accents, swap ñ for n, etc
-  var from = "àáäâèéëêìíïîòóöôùúüûñç·/_,:;";
-  var to   = "aaaaeeeeiiiioooouuuunc------";
-  for (var i=0, l=from.length ; i<l ; i++) {
+  var from = 'àáäâèéëêìíïîòóöôùúüûñç·/_,:;';
+  var to = 'aaaaeeeeiiiioooouuuunc------';
+  for (var i = 0, l = from.length; i < l; i++) {
     str = str.replace(new RegExp(from.charAt(i), 'g'), to.charAt(i));
   }
 
@@ -17,33 +38,11 @@ export function string_to_slug (str) {
     .replace(/-+/g, '-'); // collapse dashes
 }
 
-export function toggleMCE(editorId){
-  if(tinymce.get(editorId)) {
+export function toggleMCE(editorId) {
+  if (tinymce.get(editorId)) {
     tinymce.EditorManager.execCommand('mceFocus', false, editorId);
     tinymce.EditorManager.execCommand('mceRemoveEditor', true, editorId);
   } else {
     tinymce.EditorManager.execCommand('mceAddEditor', false, editorId);
   }
-}
-
-export function setTimeagoStrings(){
-  jQuery.timeago.settings.strings = {
-    prefixAgo: null,
-    prefixFromNow: null,
-    suffixAgo: "ago",
-    suffixFromNow: "from now",
-    seconds: "less than a minute",
-    minute: "a minute",
-    minutes: "%d minutes",
-    hour: "an hour",
-    hours: "%d hours",
-    day: "a day",
-    days: "%d days",
-    month: "a month",
-    months: "%d months",
-    year: "a year",
-    years: "%d years",
-    wordSeparator: " ",
-    numbers: []
-  };
 }