+++ assets/js/views/main.js
@@ -1,32 +1,31 @@
-// web/static/js/views/main.js
 /* global $, tinymce */
-import "timeago"
+import 'timeago';
 
 export default class MainView {
   mount() {
     // This will be executed when the document loads...
 
     // enable tooltips
-    $('[data-toggle="tooltip"]').tooltip()
-    $('[data-toggle="popover"]').popover()
+    $('[data-toggle="tooltip"]').tooltip();
+    $('[data-toggle="popover"]').popover();
 
-    $("time").timeago();
+    $('time').timeago();
 
     // enable base tinyMCE instance
     tinymce.init({
-  		mode : "none",
-  		branding: false,
+      mode: 'none',
+      branding: false,
       menubar: false,
       statusbar: false,
       toolbar: `styleselect forecolor | indent outdent | hr | bullist numlist | link unlink | image table | preview fullscreen`,
       plugins: `hr link lists image fullscreen preview textcolor table`
     });
 
-    $("#spinner-wrapper").addClass("invisible")
+    $('#spinner-wrapper').addClass('invisible');
   }
 
   unmount() {
     // This will be executed when the document unloads...
-    $("#spinner-wrapper").removeClass("invisible")
+    $('#spinner-wrapper').removeClass('invisible');
   }
 }