+++ lib/collaboration_web/channels/public_channel.ex
@@ -1,0 +1,7 @@
+defmodule CollaborationWeb.PublicChannel do
+  use CollaborationWeb, :channel
+
+  def join("public", _params, socket) do
+    {:ok, %{}, socket}
+  end
+end